requirements.txt
# Core ML Pipeline Dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=1.7.0
optuna>=3.0.0
mlflow>=2.5.0
evidently>=0.4.0
pydantic>=2.0.0
pyyaml>=6.0
click>=8.0.0
typer>=0.9.0
fastapi>=0.100.0
uvicorn>=0.23.0
joblib>=1.3.0
dvc>=3.0.0
requests>=2.31.0
python-multipart>=0.0.6

# Data processing
polars>=0.18.0
pyarrow>=12.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Monitoring and logging
structlog>=23.0.0
prometheus-client>=0.17.0

# Few-shot learning
transformers>=4.30.0
sentence-transformers>=2.2.0
openai>=0.27.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Distributed computing
dask[complete]>=2023.6.0
dask-ml>=2023.3.24
ray[tune]>=2.5.0
psutil>=5.9.0